requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
numpy==1.24.3
Pillow==9.5.0
redis==4.5.5
requests==2.31.0
scipy==1.10.1
six==1.16.0
spotipy==2.23.0
urllib3==2.0.3
